Extract tile store abstractions into new tileverse-tilestore module
Create tileverse-tilestore to own the format-neutral retrieval API
(TileStore) and its vector-tile specialization (VectorTileStore,
VectorTileCache, VectorTilesQuery), so future container formats
can share them. This keeps tileverse-tilematrixset and
tileverse-vectortiles focused on the tile pyramid model and the MVT
codec respectively.

VectorTileCache is now keyed on a VectorTileLoader function and a
source identifier instead of a PMTilesReader, decoupling it from
PMTiles.

Public Java package names are unchanged, downstream consumers
only need to swap the Maven coordinate.

on-behalf-of: @multiversio <info@multivers.io>
