+++ bom/pom.xml
@@ -49,6 +49,11 @@
         <version>${project.version}</version>
       </dependency>
       <dependency>
+        <groupId>io.tileverse.tilestore</groupId>
+        <artifactId>tileverse-tilestore</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
         <groupId>io.tileverse.pmtiles</groupId>
         <artifactId>tileverse-pmtiles</artifactId>
         <version>${project.version}</version>
