+++ tileverse-pmtiles/src/main/java/io/tileverse/pmtiles/store/package-info.java
@@ -13,15 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.tiling.store;
-
-import io.tileverse.tiling.matrix.Tile;
 
 /**
- * Container for a tile and its associated data.
+ * PMTiles-backed tile store implementations.
  *
- * @param <T> the type of data associated with the tile
- * @param tile the tile metadata
- * @param data the tile data
+ * <p>This package provides concrete implementations of the {@link io.tileverse.tiling.store.TileStore} and
+ * {@link io.tileverse.vectortile.store.VectorTileStore} interfaces for the PMTiles archive format.
+ *
+ * <p>Key components include:
+ *
+ * <ul>
+ *   <li>{@link io.tileverse.pmtiles.store.PMTilesVectorTileStore}: A vector tile store that reads from PMTiles files.
+ *   <li>{@link io.tileverse.pmtiles.store.PMTilesTileMatrixSet}: Factory for creating tile matrix sets matching the
+ *       specific configuration of a PMTiles archive.
+ * </ul>
  */
-public record TileData<T>(Tile tile, T data) {}
+package io.tileverse.pmtiles.store;
