+++ tileverse-tilestore/src/main/java/io/tileverse/jackson/databind/tilejson/v3/TileJSON.java
