+++ tileverse-tilestore/src/main/java/io/tileverse/tiling/store/TileData.java
@@ -15,19 +15,16 @@
  */
 package io.tileverse.tiling.store;
 
-import static java.util.Objects.requireNonNull;
+import io.tileverse.tiling.matrix.Tile;
 
-import io.tileverse.tiling.matrix.TileMatrixSet;
-
-public abstract class AbstractTileStore<T> implements TileStore<T> {
-    protected TileMatrixSet matrixSet;
-
-    protected AbstractTileStore(TileMatrixSet matrixSet) {
-        this.matrixSet = requireNonNull(matrixSet);
-    }
-
-    @Override
-    public TileMatrixSet matrixSet() {
-        return matrixSet;
-    }
-}
+/**
+ * Container for a tile's metadata and its associated content.
+ *
+ * <p>This record couples a {@link Tile} (which contains spatial metadata like extent and index) with its actual data
+ * {@code T}.
+ *
+ * @param <T> the type of data associated with the tile (e.g., {@code VectorTile} or {@code byte[]})
+ * @param tile the spatial metadata of the tile
+ * @param data the actual content of the tile
+ */
+public record TileData<T>(Tile tile, T data) {}
