+++ tileverse-tilestore/src/main/java/io/tileverse/vectortile/store/VectorTileReader.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.pmtiles.store;
+package io.tileverse.vectortile.store;
 
 import io.tileverse.tiling.store.TileData;
 import io.tileverse.vectortile.model.VectorTile;
@@ -22,8 +22,19 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Stream;
 
+/**
+ * Internal reader that extracts features from a single {@link VectorTile} based on a {@link VectorTilesQuery}.
+ *
+ * @param tileData the tile metadata and its decoded content
+ * @param query the query parameters defining which layers and features to extract
+ */
 record VectorTileReader(TileData<VectorTile> tileData, VectorTilesQuery query) {
 
+    /**
+     * Extracts all matching features from all matching layers in the tile.
+     *
+     * @return a stream of features
+     */
     public Stream<VectorTile.Layer.Feature> getFeatures() {
 
         Stream<VectorTileLayerReader> layers = extractLayers(tileData, query);
