+++ tileverse-tilestore/src/test/java/io/tileverse/jackson/databind/tilejson/v3/TilesetTypeTest.java
