+++ tileverse-tilestore/src/test/java/io/tileverse/vectortile/store/WebMercatorTransformTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.pmtiles.store;
+package io.tileverse.vectortile.store;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.within;
